+++ src/main/java/com/lirezap/pq/types/PGPing.java
@@ -17,7 +17,7 @@
  *
  */
 
-package com.lirezap.pq.enums;
+package com.lirezap.pq.types;
 
 /**
  * Postgresql C library PGPing enum.
