+++ src/main/java/com/lirezap/pq/types/PGTransactionStatusType.java
@@ -17,7 +17,7 @@
  *
  */
 
-package com.lirezap.pq.enums;
+package com.lirezap.pq.types;
 
 /**
  * Postgresql C library PGTransactionStatusType enum.
